added basic Rule class and pending specs for RuleSet class (as well as class skeleton to keep the specs running).
the Rule class represents a rule that an entry must meet. this is check with the method Rule#met_by? which takes one argument, entry. Rule is instantiated with one argument, for_field, accessible via Rule#for. When met_by? is called it should check if the passed in entry meets the rule by checking the for_field. However, the generic Rule class simply returns true for all entries. it is the job of subclasses (the actual rules) to implement a proper met_by? method.

many examples still need to be added to the RuleSet spec.