parsers can be given rulesets to check entries against. an entry must conform to each ruleset (meet each rule in each ruleset) to be in the array returned by Parser#parse_entries or the passed to block given to the same method
rulesets are added to the parser by using Parser#entries_must_conform_to(ruleset) . this method adds this ruleset to a list of rulsets it does not replace any exisiting ones. It is the client's responsibility to not add confilicting rulesets to the parser. 

Trying to keep methods alphabetical so I moved some methods around in the file.