added functionality to ruleset dsl to dynamically adding rules using a simple convention (see commit details for more info)
the ruleset method takes a block. in the block a list of rules can be created to add to your new ruleset. the rules name is listed in lower case with words seperated by underscores. The name is converted into a class under Logist::Rules (loaded dynamically if in logist/rules). For example matches is converted into the class Logist::Rules::Matches. a comma separated list of arguments can be passed to each rule definition. These arguments are the same arguments passed to the corresponding rule's initialize method. 

The example below creates an instance of RuleSet with name "My Ruleset" and 2 rules one checking for existence of a field named "field1" and another checking "field2" matches a regular expression:

ruleset :my_ruleset do 
  exists :field1
  matches :field2, /abc/
end